4-mcp/requirements.txt
langchain==0.3
langgraph==0.3
langgraph-prebuilt
langchain-openai
langchain-community
tavily-python
openai
mcp
python-dotenv==1.0.0
